Rename the Array package to Label
